pyproject.toml
[project]
name = "bookly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioredis>=2.0.1",
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "distutils",
    "fastapi[standard]>=0.117.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.25",
]

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I", "FAST"]
ignore = ["FAST002", "B008", "B904"]
